[트립트립] refactor: hook 분리
