+++ maincamp/src/app/boards/[boardId]/edit/page.tsx
@@ -14,6 +14,7 @@ export default function BoardsEditPage (){
       }
   })
 
+
     return(
       <BoardsWrite isEdit={true} data={data}/>
     );
