--- maincamp/src/app/boards/styles.module.css
@@ -1,14 +1,0 @@
-.BoardsList{
-    width: 80rem;
-    display: flex;
-    flex-direction: column;
-    /* align-items: center; */
-    gap: 1.5rem;
-
-    flex: 0 0 auto
-    
-}
-
-.BoardsListContainer{
-    width: 100%;
-}
