+++ maincamp/src/components/features/boards/boards-detail/index.tsx
@@ -20,7 +20,7 @@ export default function BoardsBoardDetailPage() {
             <Sectiontitle text={title} />
             <Writer writer={writer} date={KSTdate}/>
             <div className={styles.imagesGroup}>
-                {imagesUrl && imagesUrl.map((url: string) => <img key={url} src={url} className={styles.addimage1}/>)}
+                {imagesUrl?.map((url: string) => <img key={url} src={url} className={styles.addimage1}/>)}
             </div>
             <Sectioncontent content={contents}/>
             {GetVideoFromUrl(youtubeUrl) ? <YoutubeUrl youtubeUrl={youtubeUrl}></YoutubeUrl> : ""}
