+++ maincamp/src/components/features/boards/boards-list/index.tsx
@@ -1,0 +1,64 @@
+"use client"
+
+import DataList from "@/components/ui/list/DataList"
+import { FETCH_BOARDS_AND_COUNT } from "@/graphql/queries/board"
+import { useMutation, useQuery } from "@apollo/client"
+import styles from './styles.module.css'
+import ListRow from "@/components/ui/list/ListRow"
+import Button from "@/components/ui/button/Button"
+import { useRouter } from "next/navigation"
+import Sectiontitle from "@/components/ui/section/Sectiontitle"
+import { DELETE_BOARD } from "@/graphql/mutations/board"
+import { Fragment, MouseEvent, useState } from "react"
+import useBoardsList from "./hook"
+import Pagination from "@/components/ui/list/Pagination"
+
+
+export default function BoardsList () {
+    
+    const {changeCurrentPage, fetchBoards, fetchBoardsCount, onClickBoardsNew, flex, textAlign, pageNum, currentPage, loading } = useBoardsList();
+
+    if(loading) return ( <p> loading ... </p>)
+
+    return(
+        <div className={styles.BoardsList}>
+            <Sectiontitle text="트립토크 게시판"></Sectiontitle>
+            <Button type="submit" variant="FormBtn" onClick={onClickBoardsNew}>
+                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
+                    <path d="M5.30775 20.5C4.80258 20.5 4.375 20.325 4.025 19.975C3.675 19.625 3.5 19.1974 3.5 18.6923V5.30777C3.5 4.80261 3.675 4.37502 4.025 4.02502C4.375 3.67502 4.80258 3.50002 5.30775 3.50002H12.1598C12.4098 3.50002 12.5973 3.57852 12.7223 3.73552C12.8473 3.89269 12.9098 4.06419 12.9098 4.25002C12.9098 4.43586 12.8447 4.60736 12.7145 4.76452C12.5843 4.92152 12.3942 5.00002 12.1442 5.00002H5.30775C5.23075 5.00002 5.16025 5.03211 5.09625 5.09627C5.03208 5.16027 5 5.23077 5 5.30777V18.6923C5 18.7693 5.03208 18.8398 5.09625 18.9038C5.16025 18.9679 5.23075 19 5.30775 19H18.6923C18.7692 19 18.8398 18.9679 18.9038 18.9038C18.9679 18.8398 19 18.7693 19 18.6923V11.7923C19 11.5423 19.0785 11.3548 19.2355 11.2298C19.3927 11.1048 19.5642 11.0423 19.75 11.0423C19.9358 11.0423 20.1073 11.1048 20.2645 11.2298C20.4215 11.3548 20.5 11.5423 20.5 11.7923V18.6923C20.5 19.1974 20.325 19.625 19.975 19.975C19.625 20.325 19.1974 20.5 18.6923 20.5H5.30775ZM9.5 13.596V11.8155C9.5 11.5744 9.54683 11.3426 9.6405 11.1203C9.734 10.8978 9.86283 10.7045 10.027 10.5405L18.5598 2.00777C18.7148 1.85261 18.8853 1.73944 19.0712 1.66827C19.2571 1.59711 19.4462 1.56152 19.6385 1.56152C19.8347 1.56152 20.0231 1.59711 20.2038 1.66827C20.3846 1.73944 20.5493 1.84936 20.698 1.99802L21.9538 3.25002C22.0986 3.40519 22.2098 3.57636 22.2875 3.76352C22.365 3.95069 22.4038 4.14044 22.4038 4.33277C22.4038 4.52511 22.3708 4.71161 22.3048 4.89227C22.2388 5.07311 22.1282 5.24102 21.973 5.39602L13.4115 13.9578C13.2473 14.1218 13.0541 14.2532 12.8318 14.352C12.6093 14.4507 12.3775 14.5 12.1365 14.5H10.404C10.1462 14.5 9.93108 14.4138 9.75875 14.2413C9.58625 14.0689 9.5 13.8539 9.5 13.596ZM11 13H12.2463L18.4788 6.76727L17.8558 6.14427L17.1885 5.50202L11 11.6905V13Z" fill="white"/>
+                </svg>
+                트립토크 등록
+            </Button>
+            <div className={styles.BoardsListContainer}>
+                <DataList 
+                headerCell={["번호", "제목", "작성자", "날짜"]} 
+                flex={flex} 
+                textAlign={textAlign} 
+                // BoardsCount={fetchBoardsCount} 
+                // currentPage = {currentPage}
+                // changeCurrentPage={changeCurrentPage}
+                >
+                    {fetchBoards?.map((el: any, index: number) => {
+                        return(
+                            <Fragment key={el._id}>
+                            <ListRow
+                                num={pageNum[index]}
+                                flex={flex}
+                                textAlign={textAlign}
+                                _id={el._id}
+                                title={el.title}
+                                writer={el.writer}
+                                createdAt={el.createdAt}
+                                currentPage = {currentPage}
+                            >
+                            </ListRow>
+                            </Fragment>
+                        )
+
+                    })}
+                    <Pagination BoardsCount={fetchBoardsCount} currentPage={currentPage} changeCurrentPage={changeCurrentPage} />
+                </DataList>
+            </div>
+        </div>
+    )
+}
