+++ maincamp/src/components/features/boards/boards-write/types.ts
@@ -1,0 +1,13 @@
+export type ImageUrlArray = (string | null | undefined)[]
+
+export interface IUpdateBoardInput{
+    title?: string,
+    contents?: string,
+    youtubeUrl?: string,
+    boardAddress?:{
+        zipcode?: string,
+        address?: string,
+        addressDetail?: string,
+    },
+    images?: string[]
+}
